requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.14
django-admin-interface==0.19.1
django-admin-thumbnails==0.2.6
django-betterforms==1.2
django-colorfield==0.7.1
django-computed-property==0.3.0
django-decouple==2.1
django-environ==0.9.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-import-export==2.7.1
django-session-timeout==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
Markdown==3.3.7
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
